accessible/tests/mochitest/jsat/test_alive.html

<html>

<head>
  <title>AccessFu test for enabling</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <style type="text/css">
    .offscreen {
      position: absolute;
      left: -5000px;
      top: -5000px;
      height: 100px;
      width: 100px;
    }
  </style>

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>

  <script type="application/javascript">

    Components.utils.import("resource://gre/modules/accessibility/AccessFu.jsm");
    var gAccessFuStarted = false;

    function startAccessFu()
    {
      SpecialPowers.setIntPref("accessibility.accessfu.activate", 1);

      AccessFu.readyCallback = function readyCallback() {
        gAccessFuStarted = true;
      }
     
      AccessFu.attach(getTopChromeWindow(window));
    }

    function doTest()
    {
      startAccessFu();
      ok(gAccessFuStarted,"AccessFu initialization");
      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>

</head>

<body>

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=811307"
     title="[AccessFu] Add mochitest for enabling">
    Mozilla Bug 811307
  </a>
 
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <div>Alive!?</div>

</body>
</html>
